AppOps: Default GET_USAGE_STATS to MODE_IGNORED
The AppOp OP_GET_USAGE_STATS is defaulted with MODE_DEFAULT and this is
resolved to default to ALLOW, if the permission PACKAGE_USAGE_STATS is
requested. This can be switched off in a specific settings menu, hence
an opt-out is implemented in AOSP.

Letting 3rd parties analyze the behavior does not really add any value
for the device holder, hence an opt-in makes more sense. Usage stats
collection is now disabled by default for apps requesting that permission.

If the user wants to allow stats collection, he can enter the respective
menu in settings and allow the app to collect usage data.

Change-Id: I9e08822851cf660277e45f3023aa80d8918f45ae