Update patch set 4
Abandoned

For the moment we aren't changing this from the default, so we don't need this.

We were going to switch to cloudflare private dns by default, but testing revealed some issues after a reboot, and it'd take a while / some fiddling to get DNS working again. Not good.

In the future when we do make the change, we can revisit and bring this back as needed (although ideally in a way that doesn't involve bumping AOSP Settings version.)

Patch-set: 4
Status: abandoned
Tag: autogenerated:gerrit:abandon