Update patch set 2
Patch Set 2:

I still think it should be merged. If you look at the code, it should have been like this in the first place. There is a null/empty check for the other variable but none for this one. Why it is null to begin with should be investigated separately

Patch-set: 2
Reviewer: Gerrit User 1000002 <1000002@35266860-6252-469c-a67d-185064479fc0>