Pause apps feature
Enforce INTERACT_ACROSS_USERS to suspend apps and grant it to Launcher3 to allow it to suspend apps in other users

# This is the commit message #2:

# fixup! Pause apps feature

Prevent non-system packages from spoofing the calling package
arbitrarily, as was already prevented prior to the change this is
fixing. This helps avoid unintended consequences caused by, for
example, setting the calling package to the platform package and
attempting to add or to remove system-level suspensions, which
should not be allowed.

If the caller is from a different user than the one whose suspensions
it wants to change, ensure that it has permission to interact across
users. If it is not calling from the parent user, it must have the
full permission.

Issue: calyxos#1599
Change-Id: Ifb1bbd3cc495c6e7682908d46e41cce2fa1f5cf4

Change-Id: I96d3ffe2b0768cd9f3df1c0312f0a715087a4754
(cherry picked from commit e16dcd58bd8d87434760b7582f067bdc1e488b58)