pkg: spoof Play Store as Installer for Aurora Store
Pretend packages installed by Aurora Store have been installed by
Google Play Store (com.android.vending).

Some applications are checking that and refuse to continue if the
installer was not com.android.vending. Aurora is providing the same
service as Google Play Store regarding the installation of packages, no
reason not to consider it as a proper store.

Note: this commit is a combination of 2:

- All packages installed via Google Play Store:
  - from: Dylanger Daly <dylanger@diagnostix.io>
  - https://gitlab.com/iode/os/public/lineage/frameworks_base/-/commit/9e5c96ca4579

- Only spoof packages installed by Aurora Store and modify new API:
  - from: Vincent Vidal <vincent@vidal.pm>
  - https://gitlab.com/iode/os/public/lineage/frameworks_base/-/commit/8add3112a4a9

The main conflicts were due to the fact the original patches modified:

  services/core/java/com/android/server/pm/PackageManagerService.java

While here, the function getInstallSourceInfo() is implemented in:

  services/core/java/com/android/server/pm/ComputerEngine.java

Source:

  https://gitlab.com/iode/os/public/lineage/frameworks_base/-/commits/v3-staging

Issue: calyxos#240
Change-Id: Icec8455d372c5f328112396b86bdb3878a4a3817
Signed-off-by: Matthieu Baerts <matttbe@gmail.com>