Use profile's power button/timeout lock settings
If a managed profile has customized settings for locking after screen
timeout or for "power button instantly locks", use those for that
profile. Otherwise, use the parent's settings.

There is ordinarily no exposed UI to change these settings except for
a single work profile, and the locksettings command does not support
it, so it is not easily possible to test this without another patch
elsewhere (to be submitted).

Change-Id: I3c1e48c09dc39a2bb86448fcff1978bd16cb9ca2