Workaround: Network leaks with split-tunnel VPNs
NetworkCallback handlers that would typically cause firewall
restrictions to be updated on VPN connect/disconnect are *not* called
for split-tunnel VPNs when the system is not included. As a workaround,
we now expose the updateRestrictedModeAllowlist function of
NetworkPolicyManagerService so that it can be called directly by the
Connectivity module when there are relevant network changes.

Issue: calyxos#1081
Change-Id: Id876715c12ed44e924c165f1c13a11f56006b457