[DO NOT MERGE] Disable SquigglyProgress seekbar animation
TODO: If incorporated, it should probably be as a toggle.

Stop the media control center / notification from animating the
seekbar.

Currently has only been tested in an emulator which already failed to
actually play audio, so it would need to be verified somewhere else.

Other interesting areas:
- R.dimen.qs_media_seekbar_progress_*
  - wavelength
  - amplitude
  - phase
  - stroke_width

Could possibly accomplish this with an overlay on amplitude, but then a
toggle may not be easily possible; not sure.

Change-Id: I3aa7467b49a073e33f7d52bf96931a5dd4276f0e