Add EXTRA_DATA_REMOVED to PACKAGE_ADDED broadcast when setting an application as not hidden
The default behaviour when unhiding a package is to trigger a normal PACKAGE_ADDED broadcast which does not contain any additional information on the state of the package before the broadcast. Essentially, it is impossible to distinguish an app being unhidden (data present) from a new installation (data absent).
Modify setApplicationHiddenSettingAsUser when unhiding to add information about the previous hidden state.

Change-Id: If8ba6edc5b7c70dba8497e0fae997353ec7659df