TESTING: Debug messages for troubleshooting I84c76678
On a production device, `updateNetworksInternal` is not being called
when VPN is disconnected or reconnected. `updateSubscriptions()`
log messages should be appearing with `persist.log.tag.NetworkPolicy`
set to `VERBOSE`, but they are not, nor are `Firewall rules changed`
messages. These messages *do* appear when Wi-Fi is toggled, so the
prop is working.

If `updateNetworksInternal` is not called on VPN disconnect/reconnect,
I84c76678 may be worse than nothing for most users, since it removes
the other callbacks.

I have not been able to repro this problem on a userdebug build, so
I cannot benefit from realtime debugging. These messages are meant
to help rule out the thread getting stuck or something.

Change-Id: I00669c1e21f48b823fe94f367c080cbd9320da36