Do not ignore background vibrations from shell
For the type of bug report triggered via buttons (keychords),
dumpstate eventually drops root, taking on the shell uid.
For its initial vibration, it's still root, so there's no problem.
For the finishing vibrations, prior to this change, the vibrations
are ignored by VibrationManagerService as follows:
`Ignoring incoming vibration as process with uid= 2000 is background`

We can trust shell with this.

Change-Id: Iacab568be4b04115e06bef13148e6b0f8a066017