Delay firewall response to disconnections
Avoid a race which interferes with proper firewall policy handling.
Wait momentarily after a network disconnects before updating the
restricted mode allowlist, to avoid a race which interferes with
proper policy handling. Update the allowlist multiple times, in
case any of the prior calls were too early.

Test: Manual: Connect to an app-based VPN. Set Chromium's network
toggles such that Wi-Fi and Mobile Data are turned off, but VPN
is on. Attempt to browse; it should succeed. Disconnect the VPN.
Attempt to browse; it should fail. Prior to this, browsing was
still allowed with VPN disconnected in most cases.

Change-Id: Id677954f606c07531a48011dd4919630e99e9260