Do not add default firewall rules to rule maps
When setting a firewall rule for a specific UID, if the rule is a
default rule, delete it from our tracked rules, rather than adding
it with a value of `FIREWALL_RULE_DEFAULT`.

This mitigates a bug in which the *rules* for UIDs provided to
`setUidFirewallRulesUL` appear to be ignored, with only the UIDs
themselves being considered: so, when a UID with a DEFAULT (block)
rule is added to our map for the restricted chain, later, when the
rules are batch-updated via `setUidFirewallRulesUL`, a default rule
for the UID is included and, somewhere along the call chain,
misinterpreted as an allow rule.

TODO: Figure out where else along the call chain this should be fixed.

Change-Id: I3229fa825fb96c4fd9dff113f0e9d7a3a31fe081