Firewall: Deny access when transports unavailable
If a UID has transport-based restrictions (i.e. Mobile Data firewall
toggle off), but we are not able to determine what transports are
currently available for its use, do not give it restricted mode
access. Keep track of this so that affected UIDs have their access
eligibility reevaluated when the allowlist is next updated.

Do not cache an empty result for transports.

Change-Id: I4ec01ebd23dc457b413f44c70e2e01b8b70996e2