Prevent removal of garlic profile & restrictions
The work profiles created through garlic level selection (Safer and
Safest) are responsible for device restrictions that the user will
likely not expect to be removed when they remove the accompanying
profile. For Safest in particular, the restrictions are expected to
be maintained for the lifetime of the installation, requiring a
factory reset to remove. For this reason, we prevent the removal
of Safer and Safest profiles, except potentially from within the
Bellis app itself, which can decide whether or not to allow removal,
and if allowed, can explain the consequences to the user.

Change-Id: Icdff9b3ffb7058507dc383a0f6ed7d09fb6b9cac