Allow removal of additional work profiles
When Bellis owns the first profile, allow removal of other profiles,
which in any other situation would be removable.

When a device is organization-owned, its managed profile is always
the first profile present, and it cannot be removed except by the
profile owner app. This was designed without multiple work profiles
in mind. Preventing removal of the associated profile makes sense,
particularly because that profile may enforce device restrictions,
but other profiles should be removable.

Issue: calyxos#1518
Change-Id: Icdff9b3ffb7058507dc383a0f6ed7d09fb6b9cac