Allow network access for Lupin
Lupin (CalyxOS App Installer) does not have network access for systems
that were updated to contain it, as opposed to clean installs to the
build in which it was introduced. As a one-time fix, allow network
access for Lupin. The underlying problem that caused this is addressed
by the network access toggle implementation in I5fae5389.

Issue: calyxos#1568
Change-Id: Ia8666b5206eb2cc068f6d04576987da0c4ff9cf0