Skip firewall policies for stopped users
Adding flags to mUidOwnerMap for UIDs in stopped users takes up some
of the limited space available in that map, which can cause it to run
out of room and prevent some apps from accessing the internet. Skip
this for users or profiles that are not even running. Re-calculate
all rules when a user or profile starts or stops.

Issue: calyxos#1249
Change-Id: Icb1509893b93f729e8636ad457284e1a0b91f525