Use secondary user SID for its work profiles
When the work profile of a secondary user is set to Use One Lock
(default), ensure that its securely-stored credential is bound to
the secure user ID of the secondary user itself, rather than to
user 0. Otherwise, the work profile will only be unlockable within
a short window of when user 0 was unlocked.

Change-Id: I3d98fb6f62674dbadeb2bd8dd89a25f4679a327c