Fix Restrict USB always-allow after early disable
When Restrict USB is turned off ("always allow"), turn off USB data
before trying to turn it on. This fixes an issue in which the framework
does not enable the USB data signal when requested, presumably because
it thinks it is already on. Without this, users with Restrict USB
turned off need to set to "Never allow" and back to "Always allow"
before any USB connectivity is possible.

TODO: When done testing, if we stop turning off USB data for debuggable
devices in vendor init, then stop turning it off for them here.

Change-Id: I093d750ad46a0b693ad30a201e420a22a626934e