Fix secondary user sharing from personal to work
In ResolverActivity, to determine whether the selected profile
corresponds to the personal side as opposed to the work side, check
if its user identifier is that of the known presonal profile user ID
(i.e. the current user). Previously, this determination was based on
on whether the user was the system user, so this feature only
worked properly in the primary user / Owner.

Test: In a secondary user, create a work profile. On the personal
side, open Chromium, open any page, tap the 3-dot menu, tap Share,
tap More, and tap the Work tab. The options should show a work badge.

Issue: calyxos#1625
Change-Id: I28ee73a1ea303f416f471bd6a7deb559f24fb92e