Fix failing-to-build KeyguardViewMediatorTest
Prior to this, `atest KeyguardViewMediatorTest` does not work due to
all of our framework changes, some of which arrived in merges, causing
constructors not to match, etc.

It is probably better to fix the tests as fixups to the changes that
required the tests to be fixed in the first place, but this was hard
to determine for now due to all of the merging, etc. Also, these
changes likely fix other SystemUI-related tests, not just
KeyguardViewMediatorTest.

Possibly worth doing the above when rebasing.

Change-Id: Ib79ffe96b4910025656cfa950dd2971324c1d32e