Revert "Skip firewall rules for stopped users"
This reverts commit b9793af72fc7b4a2fe532ceb4ecd0f0b78c74f45.

Reason for revert: This is a fix for an edge case that might be breaking other edge cases. Let's see if reverting it helps. See calyxos#1799. There was also a driveby no-reply Reddit post about no connectivity in secondary users after an update that sounded like this *could* be related somehow.

Change-Id: Ifff1d9e40878e5769853caa6dbe747e8aead2449