Notify about changes to allowed transports
Add a method to NetworkPolicyCallback to be notified when a UID's
allowed transports have changed. Also include a method that can
be called to retrieve all of the latest data.

Requires Connectivity change of the same Change-Id.

Squash with: Ifeb8342698a8ba4ce2453d64dc32590aebca5fe5
"Track and send allowed transports to Connectivity"

Change-Id: Ie476f23684b00397197184e965201d6823b28de2