Prefer factory version for same version code
When a system app or APEX has been updated and exists on /data with
the same version code as the factory version (bundled with the system),
prefer the factory version, as it may contain updates that are not
reflected by its version code, as is the case for e.g. system apps
whose version codes are tied to the SDK version.

Change-Id: Ic2b1dd860446dffa02290e99323be012d0d16469