Send USER_INITIALIZE to stopped system packages
This allows the Settings app to receive USER_INITIALIZE from a new
work profile and properly complete its tasks, such as hiding itself.
Oddly, on some devices, but not others, the Settings app was starting
out in a stopped=true state in a new work profile, so IntentResolver
was filtering out its receiver (.SettingsInitialize).

Per documentation, USER_INITIALIZE can only be received by system
packages because it should be impossible for third-party apps to be
installed when the user is initialized. Therefore, we do not need
to be concerned about third-party apps being started unexpectedly.

Issue: calyxos#2113
Change-Id: I44cfbd051132a8ee4e59040966e90fb99853c8af