UsbDeviceManager: Adapt for data signaling quirks
Always treat USB as disconnected if USB data signaling is turned off.
Set USB functions to none, and send a state change event.

This allows the USB connection notification to disappear as expected
when data signaling is turned off, and it prevents the USB configuration
page in Settings from inappropriately allowing function selection
(File transfer, etc.) when USB data signaling is off.

These behaviors match the Google Pixel series, which already report
USB as disconnected when data signaling is turned off, but the Motorola
series (and maybe others that use Qualcomm's HAL) do not.

Test: Manual: Set Restrict USB to Always allow. Connect to a computer.
Observe the USB notification on device. Set Restrict USB to Deny. Check
that the notification disappears. Search Settings for USB and tap the
Connected devices > USB result. The "Use USB for" section is disabled.
Issue: calyxos#2154
Change-Id: I9d8753d4dbb18012e0d40f4991ab4c15bad14de1