Do not undo Dark Theme schedule override so often
Stop resetting the override state for Dark Theme every time a battery
broadcast is received, when automatic (sunrise/sunset) or custom
schedules are configured. Instead, only reset the overrides when the
time or timezone changes, or the begin/end time is reached. This allows
the user to turn Dark Theme off or on regardless of the schedule without
it being undone in a matter of moments.

Bug: 366644924
Issue: calyxos#2633
Change-Id: I477c50c74db244c1059541fccca0d1836cd7e415